fix(types): Achieve mypy strict compliance (0 errors)
Eliminate all 7,516+ mypy strict errors across the codebase through
systematic type annotation improvements across 13 phases.

Key changes:
- Add `-> None` to all test functions, __init__, __post_init__ methods
- Add generic type parameters (dict[str, Any], list[Any], etc.)
- Add `from __future__ import annotations` for forward references
- Fix fixture return types with Generator[T, None, None]
- Add assert narrowing for Optional access (assert x is not None)
- Use isinstance() checks for Result/Union type narrowing
- Add `# type: ignore[...]` for intentional design patterns
- Add explicit `__all__` exports for re-exported types
- Wrap Any comparisons in bool() to fix no-any-return

Files touched: 295 across agents/, bootstrap/, runtime/, protocols/

This enables transitioning CI to fully strict mypy checking.

+++ impl/claude/agents/m/tiered.py
@@ -65,7 +65,7 @@ class WorkingMemoryChunk(Generic[T]):
         """Activation adjusted by age."""
         age_minutes = (datetime.now() - self.timestamp).total_seconds() / 60
         decay = 0.95**age_minutes  # Decay per minute
-        return self.activation * decay
+        return float(self.activation * decay)
 
 
 class AttentionFilter:
